util/log: split clog.go into separate files
This change is functionality-neutral and performs the following code
structure adjustments:

- renames `logging` to `mainLog` to reflect its true nature.
- moves code away from `clog.go` to separate files:

  - `severity.go`: the `Severity` printing and conversion routines.
  - `log_entry.go`: code specific to building and parsing log entries
    (i.e. conversion between text and `log.Entry`).
  - `log_gc.go`: removing old log files (files beyond the configured
    size and number thresholds).
  - `log_flush.go`: the code that handles flushing the buffered log data
    to files and output streams.
  - `log_bridge.go`: the code that manages the coupling of the
    cockroachdb `log` package with the default Go `log` package.
  - `sync_buffer.go`: the code that manages the coupling between
    a logger and an `*os.File`.
  - `log_buffer.go`: the code that manages buffering of a single log
    entry, and optimizes the reuse of entry buffers.
  - `vmodule.go`: the code that filters the production of log entries
    depending on the source code point where the entry is produced,
    depending on `--vmodule` and `crdb_internal.set_vmodule()`
    configurations.
  - `exit_override.go`: the code that manages the callback when the
    logging system is shutting down.
  - `get_stacks.go`: the code that assembles the goroutine stack dumps
    to produce upon a log.Fatal entry.
  - `intercept.go`: the code that injects a callback for log events.
    (Used by the `logspy` debug endpoint)
  - `backtrace_trigger.go`: the code that manages
    the special purpose tracepoints.
  - `tracebacks.go`: the code that manages the the coupling with
    the standard Go environment variable `GOTRACEBACK`.

Release note: None
