+++ pkg/util/log/crash_reporting.go
@@ -206,7 +206,7 @@ func (st SafeType) WithCause(cause interface{}) SafeType {
 func ReportPanic(ctx context.Context, sv *settings.Values, r interface{}, depth int) {
 	Shout(ctx, Severity_ERROR, "a panic has occurred!")
 
-	if logging.stderrRedirected() {
+	if mainLog.stderrRedirected() {
 		// We do not use Shout() to print the panic details here, because
 		// if stderr is not redirected (e.g. when logging to file is
 		// disabled) Shout() would copy its argument to stderr
@@ -219,7 +219,7 @@ func ReportPanic(ctx context.Context, sv *settings.Values, r interface{}, depth 
 		// If stderr is not redirected, then Go's runtime will only print
 		// out the panic details to the original stderr, and we'll miss a copy
 		// in the log file. Produce it here.
-		logging.printPanicToFile(r)
+		mainLog.printPanicToFile(r)
 	}
 
 	SendCrashReport(ctx, sv, depth+1, "", []interface{}{r}, ReportTypePanic)
