+++ pkg/util/log/severity.go
@@ -1,0 +1,67 @@
+// Copyright 2019 The Cockroach Authors.
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.txt.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0, included in the file
+// licenses/APL.txt.
+
+package log
+
+import (
+	"strconv"
+	"strings"
+	"sync/atomic"
+)
+
+const severityChar = "IWEF"
+
+// get returns the value of the Severity.
+func (s *Severity) get() Severity {
+	return Severity(atomic.LoadInt32((*int32)(s)))
+}
+
+// set sets the value of the Severity.
+func (s *Severity) set(val Severity) {
+	atomic.StoreInt32((*int32)(s), int32(val))
+}
+
+// Set is part of the flag.Value interface.
+func (s *Severity) Set(value string) error {
+	var threshold Severity
+	// Is it a known name?
+	if v, ok := SeverityByName(value); ok {
+		threshold = v
+	} else {
+		v, err := strconv.Atoi(value)
+		if err != nil {
+			return err
+		}
+		threshold = Severity(v)
+	}
+	s.set(threshold)
+	return nil
+}
+
+// Name returns the string representation of the severity (i.e. ERROR, INFO).
+func (s *Severity) Name() string {
+	return s.String()
+}
+
+// SeverityByName attempts to parse the passed in string into a severity. (i.e.
+// ERROR, INFO). If it succeeds, the returned bool is set to true.
+func SeverityByName(s string) (Severity, bool) {
+	s = strings.ToUpper(s)
+	if i, ok := Severity_value[s]; ok {
+		return Severity(i), true
+	}
+	switch s {
+	case "TRUE":
+		return Severity_INFO, true
+	case "FALSE":
+		return Severity_NONE, true
+	}
+	return 0, false
+}
