+++ pkg/util/log/stderr_redirect.go
@@ -22,9 +22,16 @@ var OrigStderr = func() *os.File {
 	return os.NewFile(fd, os.Stderr.Name())
 }()
 
-// stderrRedirected returns true if and only if logging captures
-// stderr output to the log file. This is used e.g. by Shout() to
-// determine whether to report to standard error in addition to logs.
+// LoggingToStderr returns true if log messages of the given severity
+// sent to the main logger are also visible on stderr.
+func LoggingToStderr(s Severity) bool {
+	return s >= mainLog.stderrThreshold.get()
+}
+
+// stderrRedirected returns true if and only if logging to this logger
+// captures stderr output to the log file. This is used e.g. by
+// Shout() to determine whether to report to standard error in
+// addition to logs.
 func (l *loggingT) stderrRedirected() bool {
 	return l.stderrThreshold > Severity_INFO && !l.noStderrRedirect
 }
